requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langsmith==0.1.17

transformers==4.37.0
accelerate==0.26.1
sentence-transformers==2.3.1
faiss-cpu==1.7.4

huggingface-hub==0.20.3
torch==2.1.2

python-dotenv==1.0.1
